refactor(Stack): remove the interface to index file
